clang: support clang
This commit adds clang C/C++ build support. It does not attempt to
replace gcc as the linker.

It does not add clang to CI, as I don't have it installed in the
container and don't want to deal with that right now.

It removes non-lua builds from CI though, as they're redundant.

This is tested working on my machine using clang 8.0.0, unknown if
anything else works.

+++ include/nx/types.hh
@@ -8,8 +8,8 @@
 
 namespace nx {
 
-using size_t = unsigned long;
-using ssize_t = signed long;
+using size_t = uintptr_t;
+using ssize_t = intptr_t;
 
 using ptrdiff_t = signed long;
